+++ src/assets/icons/ArrowIcon.jsx
@@ -1,0 +1,14 @@
+export const ArrowIcon = ({ color, size = 14 }) => (
+    <svg width={size} height={size}   viewBox="0 0 14 14" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <g clipPath="url(#clip0_124_1620)">
+            <path d="M1.94727 4.08332C2.95588 2.33974 4.84105 1.16666 7.0002 1.16666C10.2219 1.16666 12.8335 3.77833 12.8335 6.99999C12.8335 10.2217 10.2219 12.8333 7.0002 12.8333C4.84105 12.8333 2.95588 11.6602 1.94727 9.91666" stroke={color} strokeWidth="1.2" strokeLinecap="round" strokeLinejoin="round" />
+            <path d="M7 9.33332L9.33333 6.99999L7 4.66666" stroke={color} strokeWidth="1.2" strokeLinecap="round" strokeLinejoin="round" />
+            <path d="M1.16699 7L9.33366 7" stroke={color} strokeWidth="1.2" strokeLinecap="round" strokeLinejoin="round" />
+        </g>
+        <defs>
+            <clipPath id="clip0_124_1620">
+                <rect width="14" height="14" fill="white" />
+            </clipPath>
+        </defs>
+    </svg>
+)
