+++ src/assets/icons/PlusIcon.jsx
@@ -1,0 +1,6 @@
+export const PlusIcon = ({ color, size = 18 }) => (
+
+<svg width={size} height={size} viewBox="0 0 22 22" fill="none" xmlns="http://www.w3.org/2000/svg">
+<path d="M10.9997 7.33334V14.6667M7.33301 11H14.6663M20.1663 11C20.1663 16.0626 16.0623 20.1667 10.9997 20.1667C5.93706 20.1667 1.83301 16.0626 1.83301 11C1.83301 5.9374 5.93706 1.83334 10.9997 1.83334C16.0623 1.83334 20.1663 5.9374 20.1663 11Z" stroke={color} strokeWidth="2" strokeLinecap="round" strokeLinejoin="round"/>
+    </svg>
+)
