+++ src/assets/icons/TrashIcon.jsx
@@ -1,0 +1,7 @@
+export const TrashIcon = ({ color, size = 14 }) => (
+    <svg width={size} height={size} viewBox="0 0 14 14" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <path d="M5.25 1.75H8.75" stroke={color} strokeWidth="1.2" strokeLinecap="round" strokeLinejoin="round" />
+        <path d="M1.75 3.5H12.25" stroke={color} strokeWidth="1.2" strokeLinecap="round" strokeLinejoin="round" />
+        <path d="M2.91699 3.5L3.26751 8.75775C3.31065 9.40488 3.33222 9.72844 3.40201 9.99909C3.69263 11.1261 4.60892 11.9834 5.75279 12.1984C6.02748 12.25 6.35176 12.25 7.00033 12.25V12.25C7.64889 12.25 7.97317 12.25 8.24786 12.1984C9.39174 11.9834 10.308 11.1261 10.5986 9.99909C10.6684 9.72844 10.69 9.40488 10.7331 8.75775L11.0837 3.5" stroke="#111111" strokeWidth="1.2" strokeLinecap="round" strokeLinejoin="round" />
+    </svg>
+)
