+++ src/assets/icons/index.js
@@ -4,6 +4,9 @@ export { RatingStarUserIcon } from './RatingStarUserIcon'
 export { LoginIcon } from './LogInIcon'
 export { CloseIcon } from './CloseIcon'
 export { PencilIcon } from './PencilIcon'
+export { TrashIcon } from './TrashIcon'
+export { ArrowIcon } from './ArrowIcon'
+export { PlusIcon } from './PlusIcon'
 export { RatingStarIcon } from './RatingStarIcon'
 export { ChevronDownIcon } from './ChevronDownIcon'
 export { UserCheckIcon } from './UserCheckIcon'
