+++ src/components/index.js
@@ -2,3 +2,11 @@ export * from './layouts/'
 export { Review } from './Review'
 export { Loader } from './Loader'
 export { Modal } from './Modal'
+export { Text } from './Text'
+export { AppLogo } from './AppLogo'
+export { generateSrcConfig, Image } from './Image'
+export { Button } from './buttons/Button'
+export { Input } from './fields/Input'
+export { OpacityButton } from './buttons/OpacityButton'
+export { Textarea } from './fields/Textarea'
+export { Ratings } from './fields/Ratings'
