+++ src/navigation/provider.js
@@ -2,43 +2,78 @@ import { BrowserRouter, Navigate, Route, Routes, useLocation } from 'react-route
 import { ROUTES } from './routes'
 import { Loader, MainLayout } from '../components'
 import { lazy, Suspense } from 'react'
+import { useAuthContext } from '../contexts/auth'
+import { TaskList } from '../pages/calendar/components/TaskList'
 
 const LoginPage = lazy(() => import('../pages/auth/LoginPage'))
 const RegisterPage = lazy(() => import('../pages/auth/RegisterPage'))
 const CalendarPage = lazy(() => import('../pages/calendar/CalendarPage'))
-const LandingPage = lazy(() => import('../pages/landing/LandingPage'))
 
 export const AppRouterProvider = () => (
   <BrowserRouter>
     <Routes>
       <Route path={ROUTES.HOME} element={<MainLayout />}>
         <Route index element={<div>Home Page</div>} />
-        {/* TODO: Add additional pages here */}
-      </Route>
-      <Route path={ROUTES.LOGIN} element={<AuthLayout />}>
         <Route
-          index
+          path={ROUTES.PROFILE}
+          element={
+            <ProtectedRoute>
+              <div>PROFILE</div>
+            </ProtectedRoute>
+          }
+        />
+
+        <Route
+          path={ROUTES.CALENDAR}
           element={
             <Suspense fallback={<Loader />}>
-              <LoginPage />
+              <ProtectedRoute>
+                <CalendarPage />
+              </ProtectedRoute>
             </Suspense>
           }
         />
         {/* TODO: Add additional pages here */}
-      </Route>
 
-        {/* TODO: Add 404 Page */}
-        <Route path='*' element={<div>404 Page</div>} />
-      </Routes>
-    </Suspense>
+        
+      </Route>
+<Route
+          path={ROUTES.TASKLIST}
+          element={
+              <TaskList />
+          }
+        />
+          {/* TODO: delete when calender implemente */}
+      <Route
+        path={ROUTES.LOGIN}
+        element={
+          <Suspense fallback={<Loader />}>
+            <LoginPage />
+          </Suspense>
+        }
+      />
+      <Route
+        path={ROUTES.REGISTER}
+        element={
+          <Suspense fallback={<Loader />}>
+            <RegisterPage />
+          </Suspense>
+        }
+      />
+
+      {/* TODO: Add 404 Page */}
+      <Route path='*' element={<div>404 Page</div>} />
+    </Routes>
   </BrowserRouter>
 )
 
 const ProtectedRoute = ({ children }) => {
-  const { token } = useAuthContext()
+  const { logger } = useAuthContext()
   const location = useLocation()
 
-  if (!token) return <Navigate to={ROUTES.LOGIN} replace state={{ from: location }} />
+  if (!logger) {
+    return <Navigate to={ROUTES.LOGIN} replace state={{ from: location }} />
+  }
 
   return children
 }
